real_4runr_deployment/4runr-brain/requirements.txt
# LangGraph Campaign Brain Requirements

# Core dependencies
langgraph>=0.0.40
langchain>=0.1.0
openai>=1.0.0
jinja2>=3.1.0

# Data handling
pydantic>=2.0.0
python-dateutil>=2.8.0

# 4Runr System Integration
pyairtable>=2.0.0
python-dotenv>=1.0.0

# Production dependencies
redis>=4.5.0  # For memory storage
aiofiles>=23.0.0  # For async file operations
httpx>=0.24.0  # For HTTP requests

# Development and testing
pytest>=7.0.0
pytest-asyncio>=0.21.0
black>=23.0.0
flake8>=6.0.0